Improve sign in with login
Client needs AccountId and sign in with login didn't provide that
previously.

Also email address related to sign in with login is displayed right
after initial login, so that is good to provide also. That is not
required on the other hand as it would be possible to use another
request to get that.