Simplify database connection handling
This was done because it seemed that connection closing is not working
properly as WAL mode related files did not disappear when server closed.
That however seems to be macOS bundled SQLite problem.

The macOS bundled SQLite problem was reproduced with sqlite3 CLI tool.
On Linux with the same CLI tool the WAL files disappeared when closing
a DB.