Remove public account ID
Public account ID reduces user privacy if client shows that to all
users.

The pending notification now has account IDs about the message senders.
It is better for API consistency. Originally public account IDs was
selected for that because the API authentication is different from other
API in this case, but in the end using account IDs is better.

If the former change would exist with public account IDs the location
for public API should be moved to profile server as it is only used as
public profile ID. However that causes microservice issues related to
UUID collisions which could be mitigated by making registering API to
init profile server's database for the new account. With that change
login API can not fail because of UUID collisions.