Make birthdate optional
Add storage location for user confirmation that the user is an adult.
Also save the initial profile age and the date when it was set, so it is
possible to track min and max values for the age. Updates the profile
update age check to use the new age range algorithm.